Attempted to deal with a serialization failure
As expected (after reading "Hygienic Closures for Scala Function
Serialization" If the transformation function is implemented in an
object than it works fine,  if it is outside the object (for example in
a trait) it fails.

I tried to apply the solution proposed in "Hygenic..." but I still got
an error.